fix: redirect to login when privileged session expires
