+++ app/composables/auth.js
@@ -1,8 +1,9 @@
 import { useUsersStore } from "~~/store/users";
-const userData = useUsersStore();
-const { setUserData } = userData;
+
 
 export const handleLogout = async () => {
+  const userData = useUsersStore();
+  const { setUserData } = userData;
   const { kratosUrl } = useRuntimeConfig().public;
   try {
     // Step 1: Fetch logout URL and token from the first API endpoint
@@ -45,6 +46,8 @@ export const handleLogout = async () => {
 };
 
 export const setUserDataStore = async () => {
+  const userData = useUsersStore();
+  const { setUserData } = userData;
   const { apiUrl } = useRuntimeConfig().public;
   const headers = useRequestHeaders(["cookie"]);
   try {
